+++ app/controllers/users/index.js
@@ -1,7 +1,7 @@
 /*
  Description: User management module
  */
-const db = require('../../../db_').db;
+const db = require('../../../db').db;
 const log = require('../../../log/')(module);
 const uuidv4 = require('uuid/v4');
